refactor(mcp-server): split recommend.ts into modules (SMI-1686)
Split recommend.ts (696→379 lines) into type/helper modules:
- recommend.types.ts (182 lines): Schemas, interfaces, tool definition
- recommend.helpers.ts (180 lines): Role inference, skill transformation, loading

All files now under 500 lines.
